refactor: remove media_item naming and references
